requirements.txt
aiofiles==23.2.1
aiosqlite==0.17.0
alembic==1.13.1.dev0
anyio==4.2.0
appdirs==1.4.4
argcomplete==3.1.4
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.12.3
blinker==1.7.0
Bottleneck==1.3.5
Brlapi==0.8.5
Brotli==1.1.0
build==1.0.3
CacheControl==0.14.0
certifi==2023.11.17
chardet==5.2.0
cleo==2.1.0
cliapp==1.20180812.1
click==8.1.6
colorama==0.4.6
command-not-found==0.3
contourpy==1.0.7
crashtest==0.4.1
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
databases==0.8.0
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
Django==4.2.11
django-ckeditor==6.7.0
django-environ==0.11.2
django-js-asset==2.2.0
django-widget-tweaks==1.4.8
dulwich==0.21.6
et-xmlfile==1.0.1
fastapi==0.101.0
fastimport==0.9.14
fastjsonschema==2.19.0
filelock==3.13.1
flake8==7.0.0
Flask==3.0.2
fonttools==4.46.0
fs==2.4.16
fuse-python==1.0.7
gpg==1.18.0
greenlet==3.0.3
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.4
httpx==0.26.0
idna==3.6
importlib-metadata==4.12.0
iniconfig==1.1.1
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
keyring==24.3.1
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
markdown2==2.4.11
MarkupSafe==2.1.5
matplotlib==3.6.3
mccabe==0.7.0
mdurl==0.1.2
mechanize==0.4.9
more-itertools==10.2.0
mpmath==0.0.0
msgpack==1.0.3
mutagen==1.46.0
netaddr==0.8.0
netifaces==0.11.0
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openai==1.12.0
openpyxl==3.1.2
packaging==24.0
pandas==2.1.4+dfsg
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
pipenv==2023.12.1
pipx==1.4.3
pkginfo==1.9.6
platformdirs==4.2.0
pluggy==1.4.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasyncore==1.0.2
pycairo==1.25.1
pycodestyle==2.11.1
pycryptodomex==3.20.0
pycups==2.0.1
pydantic==1.10.14
pyflakes==3.2.0
Pygments==2.17.2
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.7.0
pylev==1.4.0
pylibacl==0.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pytest==7.4.4
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-decouple==3.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
scikit-learn==1.4.1.post1
SciPy==1.11.4
seaborn==0.13.2
SecretStorage==3.3.3
setuptools==68.1.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
sqlparse==0.4.4
starlette==0.31.1
sympy==1.12
systemd-python==235
tables==3.9.2
threadpoolctl==3.1.0
toml==0.10.2
tomlkit==0.12.4
tornado==6.4
tqdm==0.0.0
trove-classifiers==2024.1.31
ttystatus==0.38
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.0+ds
wadllib==1.3.6
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.1
wheel==0.42.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
yt-dlp==2024.4.9
zipp==1.0.0
zombie-imp==0.0.2
